1. change signature of multi_img::size() 2. add copy constructor/assignment operator to vole_config 3. rework meanshift module:  - return mode vectors next to labels  - new mode-pruning code  - use more std containers  - config fix when used as a child config 4. rgb executeSOM takes somcache, not som as an optional argument

+++ core/multi_img/multi_img.cpp
@@ -574,9 +574,9 @@ void multi_img::blur(cv::Size ksize, double sigmaX, double sigmaY,
 	resetPixels();
 }
 
-unsigned int multi_img::size() const
+size_t multi_img::size() const
 {
-	return (unsigned int)bands.size();
+	return bands.size();
 }
 
 bool multi_img::empty() const
